requirements.txt
PyYAML==5.4.1
django==3.2.6
djangorestframework==3.12.4
drf-yasg==1.20.0
requests==2.26.0
uritemplate==3.0.1
